+++ libs/query/src/test-setup.ts
@@ -1,4 +1,1 @@
-import fetchMock from 'jest-fetch-mock';
 import 'jest-preset-angular/setup-jest';
-
-fetchMock.enableMocks();